작성자 확인 등등드읃으드읃 ㅇ끝!

+++ products/views.py
@@ -17,6 +17,7 @@ class ProductListAPIView(APIView):
         return Response(serializer.data)
 
     def post(self, request):
+        request.data['author'] = request.user.id
         serializer = ProductSerializer(data=request.data)
         if serializer.is_valid(raise_exception=True):
             serializer.save()
@@ -35,13 +36,18 @@ class ProductDetailAPIView(APIView):
         
     def put(self, request, id):
         product = self.get_object(id)
+        if request.user != product.author:
+            return Response(status=status.HTTP_401_UNAUTHORIZED)
         serializer = ProductSerializer(product, data=request.data, partial=True)
+        request.data['author'] = request.user.id
         if serializer.is_valid(raise_exception=True):
             serializer.save()
             return Response(serializer.data)
         
     def delete(self, request, id):
         product = self.get_object(id)
+        if request.user != product.author:
+            return Response(status=status.HTTP_401_UNAUTHORIZED)
         product.delete()
         data = {"delete": f"Product({id}) is deleted."}
         return Response(data, status=status.HTTP_200_OK)
